extract private files into own tree
