+++ private/rdf4cpp/rdf/parser/IStreamQuadIteratorSerdImpl.hpp
