+++ private/rdf4cpp/rdf/regex/RegexImpl.hpp
